+++ lib/mir-fmm/src/lib.rs
@@ -3,12 +3,12 @@ mod closures;
 mod configuration;
 mod context;
 mod declarations;
-mod definitions;
 mod entry_functions;
 mod error;
 mod expressions;
 mod foreign_declarations;
 mod foreign_definitions;
+mod function_definitions;
 mod records;
 mod reference_count;
 mod type_information;
@@ -19,11 +19,11 @@ mod yield_;
 pub use configuration::Configuration;
 use context::Context;
 use declarations::compile_declaration;
-use definitions::compile_function_definition;
 pub use error::CompileError;
 use fnv::FnvHashMap;
 use foreign_declarations::compile_foreign_declaration;
 use foreign_definitions::compile_foreign_definition;
+use function_definitions::compile_function_definition;
 use type_information::compile_type_information_global_variable;
 use yield_::compile_yield_function_declaration;
